requirements.txt
# Dependencies for Face Tracker application
opencv-python>=4.5.0
python-dotenv>=0.19.0
numpy>=1.19.0
scipy>=1.7.0